fix(app): keep Knowledge route within budget
